+++ app/controllers/pages/guidance_controller.rb
@@ -5,13 +5,13 @@ class Pages::GuidanceController < PagesController
     page_heading = session.dig(:page, :page_heading)
     guidance_markdown = session.dig(:page, :guidance_markdown)
     guidance_form = Pages::GuidanceForm.new(page_heading:, guidance_markdown:)
-    back_link = new_page_path(@form)
+    back_link = new_question_path(@form)
     render "pages/guidance", locals: view_locals(nil, guidance_form, back_link)
   end
 
   def create
     guidance_form = Pages::GuidanceForm.new(guidance_form_params)
-    back_link = new_page_path(@form)
+    back_link = new_question_path(@form)
 
     case route_to
     when :preview
@@ -19,7 +19,7 @@ class Pages::GuidanceController < PagesController
       render "pages/guidance", locals: view_locals(nil, guidance_form, back_link)
     when :save_and_continue
       if guidance_form.submit(session)
-        redirect_to new_page_path(@form)
+        redirect_to new_question_path(@form)
       else
         render "pages/guidance", locals: view_locals(nil, guidance_form, back_link), status: :unprocessable_entity
       end
@@ -31,14 +31,14 @@ class Pages::GuidanceController < PagesController
 
     guidance_form = Pages::GuidanceForm.new(page_heading: page.page_heading,
                                             guidance_markdown: page.guidance_markdown)
-    back_link = edit_page_path(@form, page)
+    back_link = edit_question_path(@form, page)
 
     render "pages/guidance", locals: view_locals(page, guidance_form, back_link)
   end
 
   def update
     guidance_form = Pages::GuidanceForm.new(guidance_form_params)
-    back_link = edit_page_path(@form, page.id)
+    back_link = edit_question_path(@form, page.id)
 
     case route_to
     when :preview
@@ -46,7 +46,7 @@ class Pages::GuidanceController < PagesController
       render "pages/guidance", locals: view_locals(page, guidance_form, back_link)
     when :save_and_continue
       if guidance_form.submit(session)
-        redirect_to edit_page_path(@form.id, page.id)
+        redirect_to edit_question_path(@form.id, page.id)
       else
         render "pages/guidance", locals: view_locals(page, guidance_form, back_link), status: :unprocessable_entity
       end
