+++ app/controllers/pages/text_settings_controller.rb
@@ -13,7 +13,7 @@ class Pages::TextSettingsController < PagesController
     @back_link_url = type_of_answer_new_path(@form)
 
     if @text_settings_form.submit(session)
-      redirect_to new_page_path(@form)
+      redirect_to new_question_path(@form)
     else
       render "pages/text_settings"
     end
@@ -35,7 +35,7 @@ class Pages::TextSettingsController < PagesController
     @back_link_url = type_of_answer_edit_path(@form)
 
     if @text_settings_form.submit(session)
-      redirect_to edit_page_path(@form)
+      redirect_to edit_question_path(@form)
     else
       render "pages/text_settings"
     end
