+++ app/views/pages/new.html.erb
@@ -1,4 +1,4 @@
-<% set_page_title(title_with_error_prefix(t("page_titles.new_page_form"), @page.errors.any?)) %>
+<% set_page_title(title_with_error_prefix(t("page_titles.new_page_form"), @question_form.errors.any?)) %>
 <% content_for :back_link, govuk_back_link_to(form_pages_path(@form)) %>
 
 <div class="govuk-grid-row">
@@ -12,7 +12,8 @@
     <%= render partial: 'form', locals: { is_new_page: true,
                                              form_object: @form,
                                              page_object: @page,
-                                             action_path: create_page_path(@form),
+                                             question_form: @question_form,
+                                             action_path: create_question_path(@form),
                                              change_answer_type_path: type_of_answer_new_path(@form),
                                              change_selections_settings_path: selections_settings_new_path(@form),
                                              change_text_settings_path: text_settings_new_path(@form),
