+++ spec/requests/pages/date_settings_controller_spec.rb
@@ -78,7 +78,7 @@ RSpec.describe Pages::DateSettingsController, type: :request do
       end
 
       it "redirects the user to the edit question page" do
-        expect(response).to redirect_to new_page_path(form.id)
+        expect(response).to redirect_to new_question_path(form.id)
       end
     end
   end
@@ -145,7 +145,7 @@ RSpec.describe Pages::DateSettingsController, type: :request do
       end
 
       it "redirects the user to the edit question page" do
-        expect(response).to redirect_to edit_page_path(form.id, page.id)
+        expect(response).to redirect_to edit_question_path(form.id, page.id)
       end
     end
