+++ spec/requests/pages/selections_settings_controller_spec.rb
@@ -66,7 +66,7 @@ RSpec.describe Pages::SelectionsSettingsController, type: :request do
       end
 
       it "redirects the user to the question details page" do
-        expect(response).to redirect_to new_page_path(form.id)
+        expect(response).to redirect_to new_question_path(form.id)
       end
     end
 
@@ -139,7 +139,7 @@ RSpec.describe Pages::SelectionsSettingsController, type: :request do
       end
 
       it "redirects the user to the question details page " do
-        expect(response).to redirect_to edit_page_path(form.id, page.id)
+        expect(response).to redirect_to edit_question_path(form.id, page.id)
       end
     end
