+++ spec/views/pages/index.html.erb_spec.rb
@@ -8,7 +8,7 @@ describe "pages/index.html.erb" do
     # mock the path helper
     allow(view).to receive(:form_path).and_return("/forms/1")
     allow(view).to receive(:type_of_answer_new_path).and_return("/forms/1/pages/new/type-of-answer")
-    allow(view).to receive(:edit_page_path).and_return("/forms/1/pages/2/edit")
+    allow(view).to receive(:edit_question_path).and_return("/forms/1/pages/2/edit/question")
     allow(view).to receive(:form_pages_path).and_return("/forms/1/pages")
     allow(view).to receive(:routing_page_path).and_return("/forms/1/new-condition")
